Reduce the bin size by increasing number of bins.
Here's a weird thing: The docs say that FFT takes a maximum number of bins of 1024. I'm currently passing 2^13, or 8192. It seems to work just fine.

I arrived at this number through experimentation. The chart mostly works with this configuration. There's still a few frequencies that produce results that are off by about 10hz.